tensorsocial-to-strapi/requirements/requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
idna==3.3
numpy==1.23.2
pandas==1.4.4
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
text-unidecode==1.3
urllib3==1.26.12
